SENSORFUSION/requirements.txt
adafruit-blinka
adafruit-circuitpython-lsm6dsox
adafruit-circuitpython-lis3mdl
pyftdi
ahrs
pygame
PyOpenGL
PyQt6
